Do not check CRC
The nitrokey embeds a wrong CRC code into the response for certain
commands. When closing the encrypted volume, for example, the CRC code
never matches the actual report data. This problem is tracked by issue
32: https://github.com/Nitrokey/nitrokey-storage-firmware/issues/32

To work around this problem for now we unfortunately have to ignore the
result of the CRC check. This change adjusts the code to do that.