Specify version requirements in a less strict manner
We try to adhere to the Semantic Versioning convention and expect
dependent crates to do the same. So far we have been fairly strict in
terms of the specific versions we consume, specifying the full version
triple (x.y.z).
This change loosens that stance by only specifying the minor version
requirement of dependencies, not the patch level (i.e., x.y).