Show error in pinentry dialog instead of printing to stdout
Currently, the error message for a wrong password is printed to the
standard output.  Yet the standard output might not be visible to the
user if they are using the curses frontend for pinentry.  Pinentry
already supports displaying an error message in the passphrase prompt.
This patch moves the error message from the standard output to the
pinentry prompt.