Port argument handling to argparse
This patch replaces the macro for argument parsing with
`argparse::ArgumentParser` from the argparse crate.  It moves the
application logic to the `commands` module and the argument parsing to
the `options` module.  An enum is used to represent the available
commands.  The code is based on the `subcommands.rs` example shipped
with argparse.