Enable rustfmt in the CI pipeline
This change enables automated code format checking by means of the
rustfmt program in the Gitlab CI pipeline. The check is performed in a
third job in the hope that this helps identify problems more easily:
users can see which of the jobs failed and focus on them in isolation,
as opposed to having a single log file or, worse, just the results of
the stages up to the first failure.

This patch resolves issue #17.