Implement the otp set subcommand
This patch implements the `otp set` subcommand that configures an OTP
slot.  There are two ways to specify an OTP secret:  as a hexadecimal
string (that means that every two characters are interpreted as a
hexadecimal representation of one byte of the secret) or as an ASCII
string (that means that the ASCII code of every character is interpreted
as one byte of the secret).  As the HOTP RFC mentions both
representations, this implementation supports both.