Implement the otp status subcommand
This patch adds the otp status subcommand that lists all OTP slots and
their current status.  To avoid hardcoding the number of slots per type,
we iterate all slots until we get an `InvalidSlot` error (assuming that
the set of valid slots is {0, ..., n} for some n).  The status command
is quite slow as we have to query each slot separately.