Implement the config set subcommand
This change implements the config set subcommand. The subcommand changes
the configuration of a Nitrokey device.

Its structure is more complex as it allows partial modifications:  The
user does not have to change all settings, but may choose to change only
some.  At the same time, the binding settings can be either set to a
value or disabled.  Therefore, we have the --{num,caps,scrol}lock
options to set a value and the --no-{num,caps,scrol}lock options to
disable the value.  If none of the two is set, the setting is not
changed.