Implement the pws get subcommand
This patch implements the pws get subcommand that provides read access
to a slot of the password safe.  Per default, all available information
– slot name, login and password – are printed.  If one or more of the
options --name, --login and --password are set, only the selected fields
are printed.  If --quiet is set, the field description is omitted such
that the output can be easily parsed by other applications.