Fix man page formatting and wording
The main change introduced in this patch is the correct usage of - and
\-.  Instead of the \(em macro for em-dashes as suggested by
man-pages(7) it uses the \(en macro:  Both dashes are typeset as a
single UTF-8 character on my terminal.  If spaces are omitted, the
resulting text is very hard to read if set in monospace fonts as the en-
or em-dash is not significantly different from a regular dash.  The
em-dash may not be used with spaces, so I chose the en-dash instead.