Set the time before generating a TOTP
This patch change the otp get command to set the Nitrokey’s time before
generating a one-time password using the TOTP command.  Per default, it
sets the time to the current system time.  If the --time option is set,
it uses its value instead.  See issue 34 [0] for a discussion of this
change.

[0] https://github.com/d-e-s-o/nitrocli/issues/34