Adjust clippy target in Gitlab CI
The clippy target as executed by the Gitlab CI excludes a bunch of lints
when performing an initial run. That is necessary because some of the
source code we rely on violates those rules and would cause the target
to fail.
The problem with the approach taken is that we list all the individual
failing lints, which quickly becomes a maintenance burden. As it turns
out clippy has the clippy::all meta-lint that subsumes all of the
explicitly specified ones and so with this change we use that instead.