Remove 'test' target from Makefile
A while ago we removed all device specific tests from the project as
part of the move to using the nitrokey crate. While adding additional
tests is a work in progress, the intention is to have them run solely by
issuing 'cargo test'.
In any case, this change removes the 'test' target from the Makefile as
it is no longer needed, because all tests can run concurrently just
fine.