Add 'nitrokey-pro' keyword to Cargo.toml
For a while now the program has supported the Nitrokey Pro device in
addition to the Nitrokey Storage. To reflect this change, this patch
adjusts the keywords in Cargo.toml to include 'nitrokey-pro' as well.
In order to not exceed the crates.io imposed limit, it removes the 'hid'
keyword.