Add the --format option to otp set to select the secret format
This patch introduces the -f, --format options for the otp set
subcommand to specify the format of the OTP secret.  Previously, the
default format as hex and ASCII format could be selected using the
--ascii option.  The new --format option takes the argument hex or
ascii, defaulting to hex, and replaces the --ascii option.

This patch does not remove the --ascii option but marks it as
deprecated.  It may not be set together with --format, and a warning is
printed if it is set.  It should be deleted with the next minor release.

This patch prepares the addition of a new format, base32.