Emit all output to custom stdio channels
This change continues and concludes the effort of using customizable
stdio channels for output of data from the program. It replaces all
existing print commands (which implicitly emit data to stdout) with
explicit write calls that output the data to the supplied Write object.
Because this object is injected from what is effectively the main
function, it will be possible for tests invoking this function to supply
custom Write objects that can buffer this data and make it available for
verification purposes.