Run unit tests as part of continuous integration
For a while now the unit tests no longer require a Nitrokey device to be
present in order to run. Right now that is because we lack device
specific tests, but even in the future we will take measures to not fail
when a device is not present (by default).
Hence, there is no reason not to run the unit tests as part of the
continuous integration pipeline. To that end, this change adds them to
said pipeline.