Reference official Gentoo ebuild from README
By now the nitrocli ebuild has been upstreamed for recent releases and
is available in the official Gentoo Portage tree.
This change updates the references to the Gentoo ebuild in the README
from the custom Github mirror to the official Gentoo Portage tree to
move users towards using the official ebuild.