Rename Command enum to Builtin
This change renames the Command enum to Builtin. With an upcoming change
we will add support for user provided functionality in the form of
extensions. Those extensions integrated with nitrocli by invocable as a
command such as status.
By renaming the Command enum to Builtin we allow for this identifier to
be used in the context of extensions in the future.