Refactor integration test code internals for program invocation
In the future we will need to perform a sequence of invocation of the
program for testing purposes, with each having a slightly different
execution context. Such a scheme does not map very well to the existing
one where we essentially just have a function invocation to run the
program. We would either have functions that produce a different
execution context or pass the in the data to modify.
Neither of these approaches is appealing and so this change reworks the
code slightly. We now can create a Nitrocli object, which contains the
data that diverges from the default execution context. This data will
eventually be modifiable by callers.